Refactor navigation structure to use AppLayout and update routing to dashboard
