+++ src/router/index.ts
@@ -6,7 +6,7 @@ const router = createRouter({
   routes: [
     {
       path: '/',
-      redirect: '/reservations'
+      redirect: '/dashboard'
     },
     {
       path: '/login',
@@ -72,7 +72,7 @@ router.beforeEach((to, _from, next) => {
   } else if (to.meta.requiresAdmin && !authStore.isAdmin) {
     next('/dashboard')
   } else if (to.path === '/login' && authStore.isAuthenticated) {
-    next('/reservations')
+    next('/dashboard')
   } else {
     next()
   }
