+++ src/views/Aircraft.vue
@@ -1,14 +1,5 @@
 <template>
-  <div class="container">
-    <header>
-      <div class="logo" @click="$router.push('/dashboard')">✈ AeroBook</div>
-      <div class="user-info">
-        <span class="role-badge" :class="`role-${authStore.userRole}`">{{ authStore.userRole }}</span>
-        <button class="btn-secondary" @click="$router.push('/dashboard')">Dashboard</button>
-        <button class="btn-secondary" @click="authStore.logout()">Logout</button>
-      </div>
-    </header>
-
+  <AppLayout>
     <div v-if="!authStore.canManageAircraft" class="alert alert-error">
       <strong>Access Denied:</strong> You don't have permission to manage aircraft. Only Admins and Operators can access this page.
     </div>
@@ -113,7 +104,7 @@
         </table>
       </div>
     </template>
-  </div>
+  </AppLayout>
 </template>
 
 <script setup lang="ts">
@@ -121,6 +112,7 @@ import { ref, onMounted } from 'vue'
 import { useAuthStore } from '../stores/auth'
 import { aircraftAPI } from '../services/api'
 import type { Aircraft } from '../types'
+import AppLayout from '../components/AppLayout.vue'
 
 const authStore = useAuthStore()
