+++ src/views/Billing.vue
@@ -1,13 +1,5 @@
 <template>
-  <div class="container">
-    <header>
-      <div class="logo" @click="$router.push('/dashboard')">✈ AeroBook</div>
-      <div class="user-info">
-        <button class="btn-secondary" @click="$router.push('/dashboard')">Dashboard</button>
-        <button class="btn-secondary" @click="authStore.logout()">Logout</button>
-      </div>
-    </header>
-
+  <AppLayout>
     <div class="page-header">
       <div>
         <h1>Billing</h1>
@@ -33,8 +25,8 @@
         <tbody>
           <tr v-for="bill in billingRecords" :key="bill.id">
             <td>#{{ bill.id }}</td>
-            <td>Member #{{ bill.member_id }}</td>
-            <td>Aircraft #{{ bill.aircraft_id }}</td>
+            <td>{{ getMemberName(bill.member_id) }}</td>
+            <td>{{ getAircraftName(bill.aircraft_id) }}</td>
             <td>{{ formatDate(bill.billing_date) }}</td>
             <td>{{ bill.tach_hours }}</td>
             <td>${{ bill.hourly_rate }}/hr</td>
@@ -56,34 +48,51 @@
         </tbody>
       </table>
     </div>
-  </div>
+  </AppLayout>
 </template>
 
 <script setup lang="ts">
 import { ref, onMounted } from 'vue'
-import { useAuthStore } from '../stores/auth'
-import { billingAPI } from '../services/api'
-import type { BillingRecord } from '../types'
+import { billingAPI, membersAPI, aircraftAPI } from '../services/api'
+import type { BillingRecord, Member, Aircraft } from '../types'
+import AppLayout from '../components/AppLayout.vue'
 
-const authStore = useAuthStore()
 const billingRecords = ref<BillingRecord[]>([])
+const members = ref<Member[]>([])
+const aircraft = ref<Aircraft[]>([])
 
-async function loadBillingRecords() {
+async function loadData() {
   try {
-    const response = await billingAPI.getAll()
-    billingRecords.value = response.data
+    const [billingRes, membersRes, aircraftRes] = await Promise.all([
+      billingAPI.getAll(),
+      membersAPI.getAll(),
+      aircraftAPI.getAll()
+    ])
+    billingRecords.value = billingRes.data
+    members.value = membersRes.data
+    aircraft.value = aircraftRes.data
   } catch (error) {
-    console.error('Error loading billing records:', error)
+    console.error('Error loading billing data:', error)
   }
 }
 
 async function markAsPaid(id: number) {
   try {
     await billingAPI.markPaid(id)
-    await loadBillingRecords()
+    await loadData()
   } catch (error) {
     console.error('Error marking as paid:', error)
   }
+}
+
+function getMemberName(memberId: number): string {
+  const member = members.value.find(m => m.id === memberId)
+  return member ? `${member.first_name} ${member.last_name}` : `Member #${memberId}`
+}
+
+function getAircraftName(aircraftId: number): string {
+  const plane = aircraft.value.find(a => a.id === aircraftId)
+  return plane ? `${plane.tail_number} - ${plane.make} ${plane.model}` : `Aircraft #${aircraftId}`
 }
 
 function formatDate(dateString: string) {
@@ -91,6 +100,6 @@ function formatDate(dateString: string) {
 }
 
 onMounted(() => {
-  loadBillingRecords()
+  loadData()
 })
 </script>
